#70 separate entities into files
