+++ server/src/main/java/com/krillsson/sysapi/core/history/db/DataType.kt
@@ -1,0 +1,5 @@
+package com.krillsson.sysapi.core.history.db
+
+enum class DataType {
+    CLOCK, VOLTAGE, PERCENT, RPM, CELCIUS, GIGABYTE
+}
