+++ server/src/main/java/com/krillsson/sysapi/persistence/PersistenceMigrator.kt
@@ -2,10 +2,10 @@ package com.krillsson.sysapi.persistence
 
 import com.codahale.metrics.MetricRegistry
 import com.krillsson.sysapi.config.SysAPIConfiguration
-import com.krillsson.sysapi.core.domain.history.HistorySystemLoadDAO
-import com.krillsson.sysapi.core.domain.history.HistorySystemLoadEntity
 import com.krillsson.sysapi.core.history.HistoryStore
 import com.krillsson.sysapi.core.history.asEntity
+import com.krillsson.sysapi.core.history.db.HistorySystemLoadDAO
+import com.krillsson.sysapi.core.history.db.HistorySystemLoadEntity
 import com.krillsson.sysapi.util.logger
 import io.dropwizard.flyway.FlywayBundle
 import io.dropwizard.hibernate.UnitOfWork
